backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "langchain>=1.0.0",
    "langchain-google-genai>=2.1.12",
    "pymongo>=4.15.3",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
